requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
django-multiselectfield==0.1.12
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
Faker==23.2.1
idna==3.6
phonenumbers==8.13.30
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
